Move runconfig blkiodev options and parsing into runconfig/opts package.
